refactor: change themes.ts to indes.ts
